+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed addSeachDate function about parsing reviewDateTime
+
 ## [3.12.4] - 2023-01-03
 
 ### Fixed