+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Removed
+- Removed store-graphql dependency
+
 ## [3.12.4] - 2023-01-03
 
 ### Fixed